Fixed a regression causing it not to compile with python 2.4
git-svn-id: http://py-leveldb.googlecode.com/svn/trunk@39 72a8a67b-58ad-ce6a-cf8f-6a8106eb947b

+++ leveldb_ext.h
@@ -36,10 +36,10 @@ typedef struct {
 	leveldb::Cache* _cache;
 
 	// number of open snapshots, associated with LevelDB object
-	Py_ssize_t n_snapshots;
+	int n_snapshots;
 
 	// number of open iterators, associated with LevelDB object
-	Py_ssize_t n_iterators;
+	int n_iterators;
 } PyLevelDB;
 
 typedef struct {
